requirements.txt
python-dotenv==1.0.1
pyotp==2.9.0
requests==2.32.3
smartapi-python==1.5.5
pycryptodome==3.21.0
matplotlib==3.8.1
logzero==1.7.0
websocket-client==1.8.0
openai==1.0.0
numpy==1.26.4

# websocket / twisted stack for SmartAPI websockets
Twisted==22.10.0
autobahn==23.6.2
service-identity==21.1.0
